ch4-ec2-get-started/requirements.txt
aws-cdk.aws-ec2==1.100.0